Small change in layout

+++ app/src/main/java/com/example/mateusz/splitwise/model/User.java
@@ -3,12 +3,12 @@ package com.example.mateusz.splitwise.model;
 public class User {
     private String username;
     private String email;
-    private int balance;
+    private double balance;
 
     public User() {
     }
 
-    public User(String username, String email, int balance) {
+    public User(String username, String email, double balance) {
         this.username = username;
         this.email = email;
         this.balance = balance;
@@ -30,11 +30,11 @@ public class User {
         this.email = email;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 }
